feat: like, comment 데이터베이스 테이블 추가
