+++ back-end/src/posts/category.service.ts
@@ -1,7 +1,8 @@
 import { Injectable, OnModuleInit } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Category } from './post.entity';
+import { Category } from './entities/category.entity';
+
 
 @Injectable()
 export class CategoryService implements OnModuleInit {
