+++ back-end/src/posts/entities/category.entity.ts
@@ -1,0 +1,14 @@
+import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from "typeorm";
+import { Post } from "./post.entity";
+
+@Entity()
+export class Category {
+  @PrimaryGeneratedColumn()
+  category_id: number;
+
+  @Column({ type: 'varchar', length: 255 })
+  name: string;
+
+  @OneToMany(() => Post, (post) => post.category)
+  posts: Post[];
+}
