+++ back-end/src/posts/entities/like.entity.ts
@@ -1,0 +1,15 @@
+import { User } from "src/user/user.entity";
+import {Entity, ManyToOne, PrimaryGeneratedColumn } from "typeorm";
+import { Post } from "./post.entity";
+
+@Entity()
+export class Like {
+    @PrimaryGeneratedColumn()
+    like_id:number;
+    
+    @ManyToOne(()=>User, {nullable:false, onDelete:'CASCADE'})
+    user:User;
+
+    @ManyToOne(()=>Post, {nullable:false, onDelete:'CASCADE'})
+    post:Post;
+}
