+++ back-end/src/posts/posts.service.ts
@@ -4,10 +4,12 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Category, Post } from './post.entity';
+
 import { Repository } from 'typeorm';
 import { User } from 'src/user/user.entity';
 import { GetPostsByCategoryDto } from './dto/category/get-posts-by-category.dto';
+import { Post } from './entities/post.entity';
+import { Category } from './entities/category.entity';
 
 @Injectable()
 export class PostsService {
@@ -165,4 +167,6 @@ export class PostsService {
       take: n,
     });
   }
+
+  
 }
